\name{output}
\alias{output}

\title{
Generating Basic Summary Statistics of Marginal Distributions
}

\description{
Producing basic summary statistics (the mean, the standard deviation and the first five modes) from the sample drawn via either the Fu-Wang algorithm or the Wang-Lee algorithm, for all marginal distributions of the target distribution.
}

\usage{
output(x)
}

\arguments{
\item{x}{
a \code{data.frame}, contains the sample drawn via either the Fu-Wang algorithm or the Wang-Lee algorithm 
}
}

\author{
Liqun Wang and Chel Hee Lee
}

