% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsample.R
\name{plot.dsample}
\alias{plot.dsample}
\title{Visualizing Wang-Lee Samples}
\usage{
\method{plot}{dsample}(x, which, ...)
}
\arguments{
\item{x}{an object produced by \code{dsample}.}

\item{which}{plot type, 1: CDF, 2: Contours, and 3: Histogram.}

\item{...}{arguments passing functions inside}
}
\description{
The samples generated by the Wang-Lee algorithm are plotted for visual examination. The plot is useful when multiple modes exist.
}
