% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dsample.R
\name{plot.dsample}
\alias{plot.dsample}
\title{Plot dsample objects}
\usage{
\method{plot}{dsample}(x, ...)
}
\arguments{
\item{x}{dsample object.}

\item{...}{arguments passing functions inside.}
}
\description{
Plot dsample objects
}

