% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dsample.R
\name{summary.samplefwl}
\alias{summary.samplefwl}
\title{Generating Basic Summary Statistics of Marginal Distributions}
\usage{
\method{summary}{samplefwl}(object, ...)
}
\arguments{
\item{object}{a \code{data.frame}, contains the sample drawn via either the Fu-Wang algorithm or the Wang-Lee algorithm}

\item{...}{dot-dot-dot arguments}
}
\description{
Producing basic summary statistics (the mean, the standard deviation and the first five modes) from the sample drawn via either the Fu-Wang algorithm or the Wang-Lee algorithm, for all marginal distributions of the target distribution.
}
\author{
Chel Hee Lee \email{chl948@mail.usask.ca}, Liqun Wang \email{liqun.wang@umanitoba.ca}
}

