% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsample.R
\name{summary.dsample}
\alias{summary.dsample}
\title{Summary Statistics of Marginal Distributions}
\usage{
\method{summary}{dsample}(object, n = 5, k = 1, ...)
}
\arguments{
\item{object}{\code{data.frame} containing the samples drawn}

\item{n}{the first n samples}

\item{k}{number of clusters}

\item{...}{arguments passing to the functions used internally}
}
\value{
\code{summary.dsample} gives a list of summary statistics.
\item{means}{Means}
\item{stdevs}{Standard deviations}
\item{modes}{Modes}
\item{hc}{object produced by \code{hclust}}
\item{grp}{cluster members produced by \code{hclust}}
\item{X}{samples generated by \code{dsample}}
\item{cdf}{cumulative distributions}
}
\description{
Producing basic summary statistics (mean, standard deviation and the first five modes) from the sample drawn for all marginal distributions.
}
